refactor: move google sheet errors
